+++ src/rend/color/error.rs
@@ -29,10 +29,17 @@ impl ColorError {
     }
 }
 
-impl crate::Error for ColorError {}
-
 mod core_impls {
     use crate::{ColorError, Display, FmtResult, Formatter};
+
+    impl crate::Error for ColorError {}
+    impl crate::ExtError for ColorError {
+        type Kind = ();
+        fn error_eq(&self, other: &Self) -> bool {
+            self == other
+        }
+        fn error_kind(&self) -> Self::Kind {}
+    }
 
     impl Display for ColorError {
         fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult<()> {