+++ src/rend/draw/error.rs
@@ -14,10 +14,17 @@ pub enum DrawError {
     DrawError,
 }
 
-impl crate::Error for DrawError {}
-
 mod core_impls {
     use crate::{Display, DrawError, FmtResult, Formatter};
+
+    impl crate::Error for DrawError {}
+    impl crate::ExtError for DrawError {
+        type Kind = ();
+        fn error_eq(&self, other: &Self) -> bool {
+            self == other
+        }
+        fn error_kind(&self) -> Self::Kind {}
+    }
 
     impl Display for DrawError {
         fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult<()> {